Use a single byte instead of Int32 for the enum

+++ src/Grid.cs
@@ -1,4 +1,4 @@
-public enum GridKind : int {
+public enum GridKind : Byte {
     Player = 0,
     Computer = 1,
     Tracker = 2,
